+++ agent.ex
@@ -1,26 +1,47 @@
 defmodule Appsignal.Agent do
-  def version, do: "91c966e"
+  def version, do: "bd1eb5f"
+
   def triples do
     %{
-      "x86_64-linux" => %{
-        checksum: "795e2d11496ada3877f6d441450ddb662380a78a28442979e77300994efe86b6",
-        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/91c966e/appsignal-x86_64-linux-all-static.tar.gz"
-       },
-      "i686-linux" => %{
-        checksum: "7f4675199a2005b8869f2a2fdf649675136bd85f9fe422b74018f90af8260232",
-        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/91c966e/appsignal-i686-linux-all-static.tar.gz"
-       },
-      "x86-linux" => %{
-        checksum: "7f4675199a2005b8869f2a2fdf649675136bd85f9fe422b74018f90af8260232",
-        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/91c966e/appsignal-i686-linux-all-static.tar.gz"
-       },
       "x86_64-darwin" => %{
-        checksum: "89f8cbbdc5291cf8bce826b0404c0af928a96bfadd8b4b636e234f1ceab85e10",
-        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/91c966e/appsignal-x86_64-darwin-all-static.tar.gz"
+        checksum: "c90ed7a396d0ab6f40c50ab0379ce66deda94e5be144dd89d146e37188b21f2b",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-x86_64-darwin-all-static.tar.gz"
        },
       "universal-darwin" => %{
-        checksum: "89f8cbbdc5291cf8bce826b0404c0af928a96bfadd8b4b636e234f1ceab85e10",
-        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/91c966e/appsignal-x86_64-darwin-all-static.tar.gz"
+        checksum: "c90ed7a396d0ab6f40c50ab0379ce66deda94e5be144dd89d146e37188b21f2b",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-x86_64-darwin-all-static.tar.gz"
+       },
+      "i686-linux" => %{
+        checksum: "a5ffc5558c6b051176a1943045f91399be4a325230e4f092ac0112ed63606a05",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-i686-linux-all-static.tar.gz"
+       },
+      "x86-linux" => %{
+        checksum: "a5ffc5558c6b051176a1943045f91399be4a325230e4f092ac0112ed63606a05",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-i686-linux-all-static.tar.gz"
+       },
+      "i686-linux-musl" => %{
+        checksum: "aaee5ccd80c8dc96401182238d39cc16f5dd9566b23e0395f316f294f08cfc73",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-i686-linux-musl-all-static.tar.gz"
+       },
+      "x86-linux-musl" => %{
+        checksum: "aaee5ccd80c8dc96401182238d39cc16f5dd9566b23e0395f316f294f08cfc73",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-i686-linux-musl-all-static.tar.gz"
+       },
+      "x86_64-linux" => %{
+        checksum: "e90af03dc243a4752137018ff92b8da2eb600c5074a841aa2d996484b9315c54",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-x86_64-linux-all-static.tar.gz"
+       },
+      "x86_64-linux-musl" => %{
+        checksum: "04b0b5abffbf6c7791866558e227044bb0b823bdef0a64c4c3d5aa99d3c482d4",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-x86_64-linux-musl-all-static.tar.gz"
+       },
+      "x86_64-freebsd" => %{
+        checksum: "b7c5f81aebf371c956d90a6ce88396925d3b571a2731081eaf9011d23569f574",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-x86_64-freebsd-all-static.tar.gz"
+       },
+      "amd64-freebsd" => %{
+        checksum: "b7c5f81aebf371c956d90a6ce88396925d3b571a2731081eaf9011d23569f574",
+        download_url: "https://appsignal-agent-releases.global.ssl.fastly.net/bd1eb5f/appsignal-x86_64-freebsd-all-static.tar.gz"
        },
     }
   end